requirements.txt
streamlit
openai
python-docx
PyPDF2
python-dotenv
Pillow
numpy
Pillow
